+++ R/demogChange.R
@@ -8,13 +8,22 @@
 #'
 #' @description
 #'
-#' Create a table of individual trees and their growth over two censuses, with many species included. 
+#' Create a table of individual trees and their growth over two censuses, with many species included.
 #'
-#' The option rnd is used to rounddown dbhs for certain intervals. The flag ctr is used to center the time variable,
-#' which is the number of years since 1960 of the interval midpoint. There is a logarithmic transformation, for which all growth<=0
-#' is converted to mingrow. There is also a power transformation, where each growth rate is raised to the power given by powertransformation.
+#' The flag ctr is used to center the time variable, which is the number of
+#' years since 1960 of the interval midpoint. There is a logarithmic
+#' transformation, for which all growth <= 0 is converted to mingrow. There is
+#' also a power transformation, where each growth rate is raised to the power
+#' given by powertransformation.
 #'
-#' In the latter, negative growths are transformed to negative, so do not need to be corrected. 
+#' In the latter, negative growths are transformed to negative, so do not need
+#' to be corrected.
+#' 
+#' @param rnd used to rounddown dbhs for certain intervals. This argument
+#'   indicates that dbhs<50 mm are rounded down to 5-mm, necessary because
+#'   saplings at BCI in 1982 and 1985 were measuring in 5-mm increments. The
+#'   growth functions in the CTFS R Package handle this.
+#'   
 #' @return 
 #' A data frame where each row gives data from one individual, including:
 #' - `tag`: the tree’s tag number;