+++ NEWS.md
@@ -1,5 +1,11 @@
 # ctfs 0.0.0.9005
 
+todo
+
+- use_package('mvtnorm'), use_package('MCMCpack'
+
+# ctfs 0.0.0.9005
+
 > How about working with the growth, mortality and recruitment functions (Demography functions)? They are already in the R Package, but need work. Use the BCI data to test and build the various documents.
 
 --Stuart